+++ tests/test_auth.py
@@ -1,7 +1,6 @@
 from __future__ import annotations
 
-from contextlib import AbstractContextManager, nullcontext
-from typing import TYPE_CHECKING
+from typing import TYPE_CHECKING, Any
 
 import pytest
 
@@ -21,55 +20,63 @@ async def test_no_token(sse_server: tuple[ServerQueue, str]) -> None:
     await queue.send_event(DUMMY_AUTH_EVENT | {SIG_HEADER: "sha256=0"})
     await queue.end_signal()
 
-    client = Monalisten(url, log_auth_warnings=True)
+    client = Monalisten(url)
     hooks_triggered = 0
+    received_issues: list[str] = []
 
     @client.on("*")
     async def _(_: types.WebhookEvent) -> None:
         nonlocal hooks_triggered
         hooks_triggered += 1
 
-    with pytest.warns(
-        UserWarning, match=f"Received {SIG_HEADER} header, but no token was set"
-    ):
-        await client.listen()
+    @client.on_internal("auth_issue")
+    async def _(_: dict[str, Any], message: str) -> None:
+        received_issues.append(message)
+
+    await client.listen()
 
     assert hooks_triggered == 2
+    assert received_issues == [f"Received {SIG_HEADER} header, but no token was set"]
 
 
 @pytest.mark.parametrize(
-    ("sig_header_entry", "warn_check", "should_be_triggered"),
+    ("sig_header_entry", "expected_issues", "should_be_triggered"),
     [
-        ({}, pytest.warns(match=f"Missing {SIG_HEADER} header"), False),
+        ({}, [f"Missing {SIG_HEADER} header"], False),
         (
             {SIG_HEADER: sign_auth_event("wrong")},
-            pytest.warns(match=f"{SIG_HEADER} header does not match set token"),
+            [f"{SIG_HEADER} header does not match set token"],
             False,
         ),
-        ({SIG_HEADER: sign_auth_event("foobar")}, nullcontext(), True),
-        ({SIG_HEADER.title(): sign_auth_event("foobar")}, nullcontext(), True),
+        ({SIG_HEADER: sign_auth_event("foobar")}, [], True),
+        ({SIG_HEADER.title(): sign_auth_event("foobar")}, [], True),
     ],
 )
 @pytest.mark.asyncio
 async def test_validation(
     sse_server: tuple[ServerQueue, str],
     sig_header_entry: dict[str, str],
-    warn_check: AbstractContextManager,
+    expected_issues: list[str],
     should_be_triggered: bool,
 ) -> None:
     queue, url = sse_server
     await queue.send_event(DUMMY_AUTH_EVENT | sig_header_entry)
     await queue.end_signal()
 
-    client = Monalisten(url, token="foobar", log_auth_warnings=True)
+    client = Monalisten(url, token="foobar")
     hook_triggered = False
+    received_issues: list[str] = []
 
     @client.on("*")
     async def _(_: types.WebhookEvent) -> None:
         nonlocal hook_triggered
         hook_triggered = True
 
-    with warn_check:
-        await client.listen()
+    @client.on_internal("auth_issue")
+    async def _(_: dict[str, Any], message: str) -> None:
+        received_issues.append(message)
+
+    await client.listen()
 
     assert hook_triggered is should_be_triggered
+    assert received_issues == expected_issues