+++ CHANGELOG.md
@@ -6,6 +6,18 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## [v0.2.0] - 2025-08-23
+
+### Added
+- Internal event hooks with `Monalisten.on_internal`
+
+### Changed
+- `MonalistenError` can now be imported directly from `monalisten`
+
+### Removed
+- `log_auth_warnings` parameter
+
+
 ## [v0.1.1] - 2025-08-20
 
 ### Fixed
@@ -22,3 +34,4 @@ Initial release 🎉
 
 [v0.1.0]: https://github.com/trag1c/monalisten/releases/tag/v0.1.0
 [v0.1.1]: https://github.com/trag1c/monalisten/compare/v0.1.0...v0.1.1
+[v0.2.0]: https://github.com/trag1c/monalisten/compare/v0.1.1...v0.2.0